refactor: convert switch and light factories to type-safe architecture
- Convert GPIOSwitchFactory to use SwitchConfig associated type
- Convert BinaryLightFactory to use LightConfig associated type
- Convert RGBLightFactory with multi-pin PWM validation
- Replace local pin validation with shared PinValidator
- Use struct value types for embedded performance

All switch and light components now follow Swift Embedded best practices.
